fix(web): move Privy sign-in flow to a layout-level controller
