+++ apps/web/src/app/api/auth/privy/complete/route.ts
@@ -50,6 +50,7 @@ export async function POST(request: Request) {
     const cookies = createAuthSessionCookieService({
       getConfig: () => getServerEnv(),
     });
+    console.info("[privy-complete] ok", parsed.data.walletAddress.slice(0, 6));
     const response = NextResponse.json({ user: result.user });
     response.cookies.set({
       name: cookies.getSessionCookieName(request),
